Re-organize data conversion stuff to keep it out of /app
